vampires-backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
django-cors-headers==4.3.1
gunicorn
psycopg2-binary
dj-database-url>=2.0

